+++ libs/adapter/nest/message/message-module.ts
@@ -1,6 +1,6 @@
 import { DynamicModule, Logger, Module } from '@nestjs/common';
 import { Context, HandlerMap } from '@pl-oss/core';
-import { MessageController } from './message.controller';
+import { MessageController } from './message-controller';
 
 @Module({ controllers: [MessageController], providers: [Logger] })
 export class MessageModule {